requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.3
python-decouple==3.0
pytz==2017.2
static3==0.7.0
whitenoise==3.3.0
